// interfaces/feed.tsx
export default interface FeedInterface {
  copyright: string;
  gamePk:    number;
  link:      string;
  metaData:  MetaData;
  gameData:  GameData;
  liveData:  LiveData;
}

export interface GameData {
  game:              Game;
  datetime:          Datetime;
  status:            Status;
  teams:             GameDataTeams;
  players:           { [key: string]: Player };
  venue:             Venue;
  officialVenue:     OfficialVenue;
  weather:           Weather;
  gameInfo:          GameInfo;
  review:            Review;
  flags:             Flags;
  alerts:            any[];
  probablePitchers:  ProbablePitchers;
  officialScorer:    OfficialScorer;
  primaryDatacaster: OfficialScorer;
}

export interface Datetime {
  dateTime:     Date;
  originalDate: Date;
  officialDate: Date;
  dayNight:     string;
  time:         string;
  ampm:         string;
}

export interface Flags {
  noHitter:            boolean;
  perfectGame:         boolean;
  awayTeamNoHitter:    boolean;
  awayTeamPerfectGame: boolean;
  homeTeamNoHitter:    boolean;
  homeTeamPerfectGame: boolean;
}

export interface Game {
  pk:              number;
  type:            AbstractGameCodeEnum;
  doubleHeader:    DoubleHeader;
  id:              string;
  gamedayType:     GamedayType;
  tiebreaker:      DoubleHeader;
  gameNumber:      number;
  calendarEventID: string;
  season:          string;
  seasonDisplay:   string;
}

export enum DoubleHeader {
  N = "N",
}

export enum GamedayType {
  C = "C",
  CF = "CF",
  Dh = "DH",
  LF = "LF",
  P = "P",
  Ph = "PH",
  RF = "RF",
  Score = "score",
  Ss = "SS",
  The1B = "1B",
  The2B = "2B",
  The3B = "3B",
}

export enum AbstractGameCodeEnum {
  A = "A",
  B = "B",
  C = "C",
  Ch = "CH",
  D = "D",
  E = "E",
  F = "F",
  FS = "FS",
  Fc = "FC",
  Ff = "FF",
  H = "H",
  Kc = "KC",
  L = "L",
  M = "M",
  R = "R",
  S = "S",
  Si = "SI",
  Sl = "SL",
  T = "T",
  TypeB = "*B",
  V = "V",
  W = "W",
  X = "X",
}

export interface GameInfo {
  attendance:          number;
  firstPitch:          Date;
  gameDurationMinutes: number;
}

export interface OfficialScorer {
  id:       number;
  fullName: string;
  link:     string;
}

export interface OfficialVenue {
  id:   number;
  link: string;
}

export interface Player {
  id:                  number;
  fullName:            string;
  link:                string;
  firstName:           string;
  lastName:            string;
  primaryNumber:       string;
  birthDate:           Date;
  currentAge:          number;
  birthCity:           string;
  birthCountry:        Country;
  height:              string;
  weight:              number;
  active:              boolean;
  primaryPosition:     Position;
  useName:             string;
  middleName?:         string;
  boxscoreName:        string;
  nickName?:           string;
  gender:              AbstractGameCodeEnum;
  isPlayer:            boolean;
  isVerified:          boolean;
  draftYear?:          number;
  pronunciation?:      string;
  mlbDebutDate:        Date;
  batSide:             BatSide;
  pitchHand:           BatSide;
  nameFirstLast:       string;
  nameSlug:            string;
  firstLastName:       string;
  lastFirstName:       string;
  lastInitName:        string;
  initLastName:        string;
  fullFMLName:         string;
  fullLFMName:         string;
  strikeZoneTop:       number;
  strikeZoneBottom:    number;
  birthStateProvince?: string;
  nameMatrilineal?:    string;
  nameTitle?:          string;
}

export interface BatSide {
  code:        AbstractGameCodeEnum;
  description: Description;
}
// ...
export interface Position {
  code:         string;
  name:         PrimaryPositionName;
  type:         PrimaryPositionType;
  abbreviation: GamedayType;
}

export enum PrimaryPositionName {
  Catcher = "Catcher",
  DesignatedHitter = "Designated Hitter",
  FirstBase = "First Base",
  Outfielder = "Outfielder",
  PinchHitter = "Pinch Hitter",
  Pitcher = "Pitcher",
  SecondBase = "Second Base",
  Shortstop = "Shortstop",
  ThirdBase = "Third Base",
}

export enum PrimaryPositionType {
  Catcher = "Catcher",
  Hitter = "Hitter",
  Infielder = "Infielder",
  Outfielder = "Outfielder",
  Pitcher = "Pitcher",
}

export interface ProbablePitchers {
  away: OfficialScorer;
  home: OfficialScorer;
}

export interface Review {
  hasChallenges: boolean;
  away:          ReviewAway;
  home:          ReviewAway;
}

export interface ReviewAway {
  used:      number;
  remaining: number;
}

export interface Status {
  abstractGameState: string;
  codedGameState:    AbstractGameCodeEnum;
  detailedState:     string;
  statusCode:        AbstractGameCodeEnum;
  startTimeTBD:      boolean;
  abstractGameCode:  AbstractGameCodeEnum;
}

export interface GameDataTeams {
  away: PurpleAway;
  home: PurpleAway;
}
// ...
export interface Venue {
  id:        number;
  name:      string;
  link:      string;
  location:  Location;
  timeZone:  TimeZone;
  fieldInfo: FieldInfo;
  active:    boolean;
}
// ...
export interface Weather {
  condition: string;
  temp:      string;
  wind:      string;
}

export interface LiveData {
  plays:     Plays;
  linescore: Linescore;
  boxscore:  Boxscore;
  decisions: Decisions;
  leaders:   Leaders;
}

export interface Boxscore {
  teams:         BoxscoreTeams;
  officials:     Official[];
  info:          NoteElement[];
  pitchingNotes: any[];
}
// ...
export interface Decisions {
  winner: OfficialScorer;
  loser:  OfficialScorer;
  save:   OfficialScorer;
}

export interface Leaders {
  hitDistance: HitDistance;
  hitSpeed:    HitDistance;
  pitchSpeed:  HitDistance;
}

export interface Linescore {
  currentInning:        number;
  currentInningOrdinal: string;
  inningState:          string;
  inningHalf:           string;
  isTopInning:          boolean;
  scheduledInnings:     number;
  innings:              Inning[];
  teams:                LinescoreTeams;
  defense:              Defense;
  offense:              Offense;
  balls:                number;
  strikes:              number;
  outs:                 number;
}
// ...
export interface Plays {
  allPlays:      AllPlay[];
  currentPlay:   CurrentPlay;
  scoringPlays:  number[];
  playsByInning: PlaysByInning[];
}
// ...
export interface Count {
  balls:   number;
  strikes: number;
  outs:    number;
}
// ...
export interface AllPlayPlayEvent {
  details:            PlayEventDetails;
  count:              Count;
  index:              number;
  startTime:          Date;
  endTime?:           Date;
  isPitch:            boolean;
  type:               PlayEventType;
  player?:            OfficialVenue;
  pitchData?:         PitchData;
  playId?:            string;
  pitchNumber?:       number;
  hitData?:           HitData;
  actionPlayId?:      string;
  isBaseRunningPlay?: boolean;
  isSubstitution?:    boolean;
  position?:          Position;
  battingOrder?:      string;
  replacedPlayer?:    OfficialVenue;
}

export interface PlayEventDetails {
  description:    string;
  event?:         string;
  eventType?:     string;
  awayScore?:     number;
  homeScore?:     number;
  isScoringPlay?: boolean;
  hasReview:      boolean;
  call?:          BatSide;
  code?:          string;
  ballColor?:     BallColor;
  trailColor?:    TrailColor;
  isInPlay?:      boolean;
  isStrike?:      boolean;
  isBall?:        boolean;
  type?:          BatSide;
  fromCatcher?:   boolean;
  runnerGoing?:   boolean;
}

export enum BallColor {
  RGBA170211110 = "rgba(170, 21, 11, 1.0)",
  RGBA268619010 = "rgba(26, 86, 190, 1.0)",
  RGBA391613910 = "rgba(39, 161, 39, 1.0)",
}

export enum TrailColor {
  RGBA0025410 = "rgba(0, 0, 254, 1.0)",
  RGBA08525410 = "rgba(0, 85, 254, 1.0)",
  RGBA119015210 = "rgba(119, 0, 152, 1.0)",
  RGBA152010110 = "rgba(152, 0, 101, 1.0)",
  RGBA153171010 = "rgba(153, 171, 0, 1.0)",
  RGBA18803310 = "rgba(188, 0, 33, 1.0)",
  RGBA50022110 = "rgba(50, 0, 221, 1.0)",
}

export interface HitData {
  launchSpeed:   number;
  launchAngle:   number;
  totalDistance: number;
  trajectory:    Trajectory;
  hardness:      Hardness;
  location:      string;
  coordinates:   HitDataCoordinates;
}

export interface HitDataCoordinates {
  coordX: number;
  coordY: number;
}

export enum Hardness {
  Hard = "hard",
  Medium = "medium",
  Soft = "soft",
}

export enum Trajectory {
  FlyBall = "fly_ball",
  GroundBall = "ground_ball",
  LineDrive = "line_drive",
  Popup = "popup",
}

export interface PitchData {
  startSpeed?:      number;
  endSpeed?:        number;
  strikeZoneTop:    number;
  strikeZoneBottom: number;
  coordinates:      { [key: string]: number };
  breaks:           Breaks;
  zone?:            number;
  typeConfidence?:  number;
  plateTime?:       number;
  extension?:       number;
}

export interface Breaks {
  breakAngle?:    number;
  breakLength?:   number;
  breakY?:        number;
  spinRate?:      number;
  spinDirection?: number;
}

export enum PlayEventType {
  Action = "action",
  Pickoff = "pickoff",
  Pitch = "pitch",
}
// ...
export interface MetaData {
  wait:          number;
  timeStamp:     string;
  gameEvents:    string[];
  logicalEvents: string[];
}